requirements.txt
hydra-core
pytorch-lightning
numpy
pytorch
torchinfo
xarray
scikit-learn